Ritual procedure small update
Check the card's original type to be a Ritual Monster rather than its current type. Allows Ritual Summoning effects to work with a Ritual Monster that's currently in the Spell/Trap Zone if that Ritual Summoning effect specifically includes the Spell/Trap Zone in the locations where it can summon a Ritual Monster from.