"Cupid Pitch" fixes
Fixed its Level changing effect in the following scenarios:
- When using a monster that changes its Level for the Synchro Summon as material (e.g. "Road Synchron", it should use that changed Level).
- When using a non-Tuner monster that can be treated as a Tuner (e.g. the Token summoned by "Ringowurm the Hundred Apple Dragon").
- When using multiple Tuner monsters that can be treated as non-Tuners (e.g. "Phantom King Hydride").
In cases where there are multiple options for what could be the Tuner for the summon, the player will be prompted to pick which monster they want to use as the Tuner.

Also, its last effect shouldn't do damage if the Synchro Summoned monster is no longer face-up on the field on resolution.