"Assault Synchron" fix
Updated its 2nd effect to use "EVENT_LEAVE_FIELD" instead so that it can properly exclude itself from the event group. Fixes issues where the effect would be usable if "Assault Synchron" was sent to the GY at the same time that a Dragon Synchro Monster was Tributed or banished (it needs to already be in the GY to see it happening).