Fixes for cards that equip or attach a monster they destroyed by battle to themselves
For the following cards:
- Predaplant Moray Nepenthes
- Infinitrack River Stormer
- Infinitrack Mountain Smasher
- Infinitrack Earth Slicer
Shouldn't be able to use their respective effects if the destroyed card isn't a monster after being destroyed (e.g. it was Trap Monster). The "Infinitrack" monsters should attach the monster even if they're face-down on resolution. The effects should be considered effects that move a card from the GY only if the destroyed monster ended up in the GY.