Fix for some Continuous Traps that target a card(s) on the field
They should be able to target themselves even if their effect is used in the same Chain Link as the card's activation, but they shouldn't be considered to be face-up on the field for the activation check of said effect. For example, if there is no other face-up Spell/Trap on the field then you cannot use the effect of "Unending Nightmare" in the same Chain Link as the card's activation, but if there is another face-up Spell/Trap on the field then you can, and in that case you can even target "Unending Nightmare" itself.

Updated the scripts of the following cards:
- Unending Nightmare
- Virtual World Gate - Chuche